Alteracoes feitas no codigo do arquivo gerador.py separando suas clases em outros arquivos.
