+++ harpia/connection.py
@@ -1,0 +1,7 @@
+#####################################CLASSES############################
+##################### connection ###########################
+class connection:
+    sourceOutput = '00'
+    destinationNumber = '00'
+    destinationInput = '00'
+    connType = " "
